move metrics AST things into dedicated file
